+++ app/src/test/java/app/knotwork/android/domain/engine/PipelineSecurityContourTest.kt
@@ -14,6 +14,7 @@ import app.knotwork.android.domain.engine.executors.InputNodeExecutor
 import app.knotwork.android.domain.engine.executors.LiteRtNodeExecutor
 import app.knotwork.android.domain.engine.executors.NodeExecutorFactory
 import app.knotwork.android.domain.engine.executors.OutputNodeExecutor
+import app.knotwork.android.domain.engine.executors.PipelineNodeExecutor
 import app.knotwork.android.domain.engine.executors.QueueProcessorNodeExecutor
 import app.knotwork.android.domain.engine.executors.SummaryNodeExecutor
 import app.knotwork.android.domain.engine.executors.SystemNodeExecutor
@@ -63,6 +64,7 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.rules.TemporaryFolder
 import java.io.File
+import javax.inject.Provider
 
 /**
  * Cross-cutting security-contour test: drives the security guards of the
@@ -191,6 +193,7 @@ class PipelineSecurityContourTest {
                 pendingInteractionRepository,
                 clarificationNotifier,
             ),
+            PipelineNodeExecutor(mockk(relaxed = true), mockk(relaxed = true), Provider { mockk(relaxed = true) }),
         )
 
         engine = GraphExecutionEngine(